requirements_core.txt
# Core dependencies only - for stable installation

# YOLOv8 and deep learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Visualization and progress
matplotlib>=3.7.0
tqdm>=4.65.0

# Data processing
pandas>=2.0.0

